requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
cmake==4.1.0
colorama==0.4.6
face_recognition_models==0.3.0
Flask==3.1.1
Flask-WTF==1.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.6
opencv-python==4.12.0.88
pillow==11.3.0
Werkzeug==3.1.3
WTForms==3.2.1
